requirements.txt
```
Django==5.2.8
djangorestframework==3.15.2
